+++ test/TaskSetTest.js
@@ -3,6 +3,7 @@ Object.defineProperty(exports, "__esModule", { value: true });
 var chai = require("chai");
 var Q = require("q");
 var Defer = require("ts-promises/Defer");
+var Tasks = require("../task/Tasks");
 var TaskSet = require("../task/TaskSet");
 var TaskState = require("../task/TaskState");
 var asr = chai.assert;
@@ -31,6 +32,35 @@ suite("TaskSet", function TaskSetTest() {
         }
         return tasks;
     }
+    test("README-example", function README_exampleTest(done) {
+        var taskSet = new TaskSet(null, function (name) { return console.log("success:", name); }, function (name) { return console.log("failure:", name); });
+        taskSet.startTask("task-1", Tasks.newTask("a", function () { return "result a"; }).start());
+        taskSet.startTask("task-2", Q.resolve("result b"));
+        taskSet.startTask("task-3", Q.reject("error c"));
+        Q.all(taskSet.getPromises())
+            .then(function (results) { console.log("done:", results); done(); }, function (err) { console.error("error:", err); done(); });
+    });
+    test("task-set-function-and-promise-task-mix", function taskSetFunctionAndPromiseTaskMixTest(done) {
+        var taskSet = new TaskSet(null, null, null);
+        taskSet.startTask("task-1", function () { return "result a"; });
+        taskSet.startTask("task-2", Tasks.newTask("b", function () { return "result b"; }).start());
+        taskSet.startTask("task-3", Q.resolve("result c"));
+        taskSet.startTask("task-4", Q.reject("error d"));
+        asr.equal(taskSet.getTasks().size, 4);
+        asr.equal(taskSet.getPromises().length, taskSet.getTasks().size);
+        Q.all(taskSet.getPromises()).done(function (results) {
+            done("unexpected success");
+        }, function (err) {
+            Q.all(taskSet.getPromises()).done(function (results) {
+                var allResults = taskSet.getCompletedTasks().map(function (t) { return t.task.getResult(); });
+                asr.deepEqual(allResults.sort(), ["result a", "result b", "result c"]);
+                asr.equal(err, "error d");
+                done();
+            }, function (err2) {
+                done("unexpected 2nd error");
+            });
+        });
+    });
     test("task-set-success", function taskSetSuccessTest(done) {
         // test success
         var taskSet = new TaskSet();
@@ -44,8 +74,7 @@ suite("TaskSet", function TaskSetTest() {
             asr.equal(task2.state, TaskState.COMPLETED);
             done();
         }, function (err) {
-            asr.equal(true, false, "unexpected error");
-            done();
+            done("unexpected error");
         });
     });
     test("task-set-failure", function taskSetFailureTest(done) {
@@ -56,10 +85,9 @@ suite("TaskSet", function TaskSetTest() {
         var task3 = taskSet.startTask("task-err-1", createTaskErr1("error-1"));
         var task4 = taskSet.startTask("task-err-2", createTaskErr2("error-2", 10));
         Defer.when(taskSet.getPromises()).done(function (res) {
-            asr.equal(true, false, "unexpected success");
-            done();
+            done("unexpected success");
         }, function (err) {
-            asr.equal(true, err == "error-1" || err == "error-2");
+            asr.isTrue(err == "error-1" || err == "error-2");
             done();
         });
     });
@@ -83,8 +111,7 @@ suite("TaskSet", function TaskSetTest() {
             taskSet.clearCompletedTasks();
             done();
         }, function (err) {
-            asr.equal(false, true, "unexpected error");
-            done();
+            done("unexpected error");
         });
     });
 });